more fiddlin'

+++ data/gwas_clean.py
@@ -8,17 +8,17 @@ data = pd.read_csv(filepath, sep='\t', low_memory=False)    # Reads gwas tsv
 data = data.loc[data['DISEASE/TRAIT']=='Type 1 diabetes']   # Select only rows regarding type 1 diabetes
 data = data.loc[data['CHR_ID']=='6']                        # Select only rows for chromosome 6
 data = data[["SNPS","REGION","CHR_POS","P-VALUE"]] # maybe also include STRONGEST SNP-RISK ALLELE and RISK ALLELE FREQUENCY ?
-data.reset_index(drop=True,inplace=True)    # Resets index back to 0
 
-rename_dict={}
+renameDict={}
 for col in data.columns:
     newCol = re.sub(r'\W+', '_', col)   # Replaces special characters and whitespace with underscores
-    rename_dict.update({col:newCol})
-data=data.rename(columns=rename_dict)
+    renameDict.update({col:newCol})
+data=data.rename(columns=renameDict)
 
 data=removeDupes(data)      # Remove duplicates (leaving the entry with largest p value)
 
 filepath = getPath(fileOut)
-os.remove(filepath)
+if os.path.exists(filepath):        # If the file exists,
+    os.remove(filepath)             # delete it.
 data.to_csv(filepath, sep='\t')
 print("\ndone\n")
